+++ Tests/AdvancedOperationTests/Operations/AsynchronousBlockOperationTests.swift
@@ -31,24 +31,6 @@ final class AsynchronousBlockOperationTests: XCTestCase {
     #endif
   }
 
-  // TODO
-//  func testCancelledExecution() {
-//    let operation = AsynchronousBlockOperation { complete in
-//      DispatchQueue(label: "\(identifier).\(#function)", attributes: .concurrent).asyncAfter(deadline: .now() + 2) {
-//        complete()
-//      }
-//    }
-//    XCTAssertTrue(operation.isAsynchronous)
-//    XCTAssertTrue(operation.isConcurrent)
-//    let expectation1 = XCTKVOExpectation(keyPath: #keyPath(Operation.isFinished), object: operation, expectedValue: true)
-//    operation.installLogger()
-//    operation.start()
-//    operation.cancel()
-//    wait(for: [expectation1], timeout: 4)
-//    XCTAssertTrue(operation.isCancelled)
-//    XCTAssertEqual(operation.name, "AsynchronousBlockOperation")
-//  }
-
   func testCancelledExecutionBeforeStarting() {
     let operation = AsynchronousBlockOperation { complete in
       DispatchQueue(label: "\(identifier).\(#function)", attributes: .concurrent).asyncAfter(deadline: .now() + 2) {
@@ -63,16 +45,6 @@ final class AsynchronousBlockOperationTests: XCTestCase {
     wait(for: [expectation1], timeout: 4)
     XCTAssertTrue(operation.isCancelled)
   }
-
-    // TODO
-//  func testEarlyBailOut() {
-//    let operation = AsynchronousBlockOperation { $0() }
-//    let expectation1 = XCTKVOExpectation(keyPath: #keyPath(Operation.isFinished), object: operation, expectedValue: true)
-//    operation.cancel()
-//    operation.start()
-//    wait(for: [expectation1], timeout: 4)
-//    XCTAssertTrue(operation.isCancelled)
-//  }
 
   func testBlockOperationCompletedInAsyncQueue() {
     let operation = AsynchronousBlockOperation { complete in